backend/requirements.txt
Django==5.2.12
djangorestframework==3.16.1
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.1
gunicorn==25.1.0
gevent==24.11.1
whitenoise[brotli]==6.11.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
requests==2.32.5
beautifulsoup4==4.14.3
lxml==6.0.2
tldextract==5.3.1
dnspython==2.7.0
pyOpenSSL==25.0.0
cryptography==44.0.2
python-whois==0.9.6
jsbeautifier==1.15.1
celery[redis]==5.4.0
redis==5.2.0
drf-spectacular==0.29.0
